refactor && fix drop bug when not area
